Rewrite to allow arbitrary style check commands
Although the only one currently supported is hardcoded, the
infrastructure is now in place to use an arbitrary command for style
checking modified files.

This should make the tool useful for any programming language, and make
it easy for users to pass arbitrary config options.

Now I just need to generalize the config options out to a config file or
similar...