Add a basic config file
It is now possible (and necessary) for end users to use it to specify
the style check command, and the regex to pull the line number a line of
output from the command applies to.

It's just a basic JSON file, and the way we load it is dumb, but it works.